Added unified Route
First task
