+++ server.js
@@ -11,7 +11,7 @@ const methodOverride = require('method-override')
 const Event = require('./models/Event')
 const Users = require('./models/Users')
 // activate morgan in order to get an idea of the get and post requests which are being send to
-// const morgan = require('morgan');
+const morgan = require('morgan');
 
 const app = express()
 const port = process.env.PORT || 5000
@@ -71,7 +71,7 @@ app.get('/admin/', (req, res) => {
   res.render('adminLogin')
 })
 
-// app.use(morgan('tiny'));
+app.use(morgan('tiny'));
 
 app.use('/users', userRouter)
 app.use('/gform', gformRouter)
@@ -115,8 +115,8 @@ app.post('/users/profile/image/upload/:id', upload.single('file'), (req, res) =>
   // res.json({file:req.file});
 })
 
-// returns an image stream to show as prof pic    || todo add it to the ejs once the server is made online
-app.get('image/:filename', (req, res) => {
+// returns an image stream to show as prof pic  
+app.get('/image/:filename', (req, res) => {
   const file = gfs
     .find({
       filename: req.params.filename
@@ -133,7 +133,7 @@ app.get('image/:filename', (req, res) => {
 
 // delete request as per documentation to clear all chunks probably need to preserve the object id
 // implement only if required based on front end design. Since it requires and additional param preserved
-app.post('image/del/:img', (req, res) => {
+app.post('/image/del/:img', (req, res) => {
   gfs.delete(new mongoose.Types.ObjectId(req.params.img), (err, data) => {
     if (err) return res.status(404).json({ err: err.message })
     res.status(200)
@@ -180,4 +180,20 @@ app.post('/users/add_event/:club_head_id/save', upload.single('poster'), (req, r
 
 app.get('/test',(req,res)=>{
   res.render('contact')
+})
+
+// use the following format for update requests in this route
+app.post('/users/profile/:id', upload.single('profpic'), (req, res) => {
+  const id = req.body.id
+  const uid = req.params.id
+  const change = {
+    pswd: req.body.pswd,
+    name: req.body.name,
+    contact: req.body.contact,
+    email_id: req.body.email_id
+  }
+  Users.findByIdAndUpdate(id, change)
+    .then((user) => {
+      res.redirect('/users/profile/'+uid)
+    })
 })
